app/services/gpt-context/requirements.txt
langchain==0.0.168
chromadb==0.3.22
python-dotenv==1.0.0
gptcache==0.1.23
openai==0.27.6
tiktoken==0.4.0